+++ features/leave/components/skeletons/dashboard-skeleton.tsx
@@ -1,5 +1,5 @@
-import { Card, CardContent, CardHeader } from '@/components/ui/card';
-import { Skeleton } from '@/components/ui/skeleton';
+import { Card, CardContent, CardHeader } from '@/components/shadcn/card';
+import { Skeleton } from '@/components/shadcn/skeleton';
 
 export function DashboardSkeleton() {
   return (
